+++ src/main/scala/org/podval/xml/Html.scala
@@ -3,8 +3,8 @@ package org.podval.xml
 import zio.blocks.chunk.Chunk
 import zio.blocks.html.Dom as XML
 
-// XML AST for ZIO Blocks HTML
-given Html: XmlAst[XML.Element]:
+// XML AST for ZIO Blocks HTML. Not a package given: `import org.podval.xml.html.given`.
+object Html extends XmlAst[XML.Element]:
   override type Node = XML
 
   override def text(text: String): Node = XML.text(text)