+++ src/test/scala/org/podval/xml/XmlCodecSpec.scala
@@ -1,4 +1,6 @@
 package org.podval.xml
+
+import org.podval.xml.scalaxml.given
 import org.scalatest.funsuite.AnyFunSuite
 import zio.blocks.schema.{Modifier, Schema}